% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managephase.R
\name{contingencise}
\alias{contingencise}
\title{Assign Contingency of Experimental Scheme}
\usage{
contingencise(a, metadata = digcondition())
}
\arguments{
\item{a}{data frame}

\item{metadata}{named vector
\describe{
\item{values}{numeric values of reward probabilities}
\item{names()}{character vector of labels present in data}
}
Default: call \code{digcondition}}
}
\value{
dataset with added columns:
\describe{
\item{rp}{numeric. reward probability}
\item{contingency}{integer. number of experimental stage (=contingency)}
}
}
\description{
\code{contingencise()} assigns a contingency mark
(number of experimntal stage) to every visit in dataset
on the basis on reward probabilities.
}
\examples{

dx <- contingencise(dx)

}
