% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managephase.R
\name{digcondition}
\alias{digcondition}
\title{Provide default coding for corner condition}
\usage{
digcondition()
}
\value{
a named atomic vector
\describe{
\item{values}{numeric values of reward probabilities}
\item{names()}{character vector of labels present in data}
}
}
\description{
Default values passed to \code{contingencise()}
}
