% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{digform}
\alias{digform}
\title{Provide Formatting Standards On Call}
\usage{
digform(version)
}
\arguments{
\item{version}{Character or integer. Version of software:
\describe{
\item{#1}{"new", "star", "newer", 1}
\item{#2}{"old", "plus", "older, "proper", 2}
}}
}
\value{
a named list
}
\description{
\code{digform()} returns a list of variables specifying
experimental files formatting,
which varies dependenting on software version
}
\details{
It's called also inside other package's functions.
}
\examples{
digform(1)

}
