% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking.R
\name{hint}
\alias{hint}
\title{What Did I Set While Recording}
\usage{
hint(A)
}
\arguments{
\item{A}{data frame}
}
\value{
list
\describe{
\item{conditiondevice}{number of visits performed depending
on condition label and device used}
\item{cornercondition}{number of visits performed split
by corner-condition assignment}
\item{corneropen}{number of visits while door got opened (
access granted)}
\item{hint}{Fraction on real access being granted under
all the condition labels}
\item{outlier}{number of visits split by participants and
condition label; it's helping to spot outlier}
}
}
\description{
\code{hint()} prints a list with basic information
about the dataset formatting
}
\examples{
dx <- hint(dx)
}
