% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icager.R
\docType{package}
\name{icager}
\alias{icager}
\title{icager: Preprocess Inteligent System's Experimental Data Easily}
\description{
The package enables painless, quick, automatic import and
preprocessing of experimental data from a standard
intelligent housing system, intended to test cognitive
abilities on group-housed mice under various
neuropsychological factors.
}
\details{
For tutorial call \code{utils::browseVignettes(icager)}
}
