% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{import}
\alias{import}
\title{Preprocess Experiment's Files}
\usage{
import(mydirs, version)
}
\arguments{
\item{mydirs}{Character vector. Paths to files}

\item{version}{Character or integer. Version of software:
\describe{
\item{"old", alteratively: "plus", "older, "proper", 2}
\item{"new", alteratively: "star", "newer", 1}
}
For details call \code{?digform(version)}}
}
\value{
list [data frame]
}
\description{
\code{import} imports all experimentals files
stored in given directories, parses on the basis on
prespecified standards and returns a list
}
\details{
Warning: The output object (list of data frames)
can be large in size. It depends entirely on number
of files provided. Loading one experiment dataset
per time is recomended.
}
\examples{
mydata <- import(
  list.files(
   path = "F://phd_data/DATA/", #or: choose.dir() WIN
   all.files = T,
   recursive = F,
   full.names = T) \%>\%
 .[grep(., pattern = "/2019-", fixed = T)],
 version = 2) \%>\%
dplyr::bind_rows() #or: rbind()

}
