% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfile.R
\name{importdoor}
\alias{importdoor}
\title{Get The Door Openings Info}
\usage{
importdoor(version, target = file.choose())
}
\arguments{
\item{version}{Numeric. Indicates type of
software used; for more info call \code{digformat(data)}}

\item{target}{Character. Path to file; Default
calls widget to choose a file WIN by \code{file.choose()}}
}
\value{
Data frame with 3 variables:
\describe{
  \item{TimeDoorOpened}{exact time of door opening; <dttm>}
  \item{deviceid}{name of the house cage; <chr>}
  \item{Corner}{where the door got opened; <int>}
}
}
\description{
\code{importfile} reads a single file word by word on,
selects proper informations on the basis of standards,
parses and returns a data table
}
\details{
Files containing all hardware info usually are clumsy,
so here the needed variables are get by position
rather than by name. Please note: A door opening is equal
to access to a reward.
}
\examples{
mydoorinfo <- importdoor(1)
}
