% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managephase.R
\name{plotit}
\alias{plotit}
\title{Visualise Experimental Scheme}
\usage{
plotit(A)
}
\arguments{
\item{A}{a dataset}
}
\description{
\code{plotit()} presents the actual
experimental scheme with contingencies
durations in a default form
}
\examples{
plotit(contingencise(dx))

}
