% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managephase.R
\name{plotscheme}
\alias{plotscheme}
\title{Visualise Experimental Scheme}
\usage{
plotscheme(A, ...)
}
\arguments{
\item{A}{data frame. a dataset or experimental scheme (see: \code{printscheme()})}
}
\description{
\code{schematize()} shows colorful abstract representation of
experimental scheme
}
\examples{
dx <- contingencise(dx) \%>\% schematize()

}
