% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managephase.R
\name{printscheme}
\alias{printscheme}
\title{Provide Experimental Scheme}
\usage{
printscheme(A, sorted = T, sep = "x")
}
\arguments{
\item{A}{data frame}

\item{sorted}{logical. indicates whether a contingency
label should consist of sorted values}

\item{sep}{character. a separator used in a contingency label}
}
\value{
data frame
\describe{
\item{contingency}{number of contingency <int>}
\item{`1/2/3/4`}{corners reward proabilities <dbl>}
\item{start}{time of first visit made under the contingency <dttm>}
\item{end}{time od last visit made under the contingency <dttm>}
\item{duration}{difference in time between start and end <difftime>}
\item{label}{label of the contingency}
}
}
\description{
\code{printscheme()} returns a data frame of concurrent
experimental contingencies with variables describing them.
}
\details{
An experimental task consists of subsequent changes
of reward probabilisty to corner assignment (contingency).
Here the total set of data-driven contingencies are stored
in a single data frame.
}
\examples{

printscheme(dx)

}
