% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managephase.R
\name{schematize}
\alias{schematize}
\title{Assign Experimental Module}
\usage{
schematize(A, info = NA, help = F)
}
\arguments{
\item{A}{data frame}

\item{info}{data frame. Character vector of labels
assigned to contingency (also known as module name to experimental phase)
\describe{
\item{continency}{character vector or integer. Label of contingency}
\item{info}{character vector. Label of module}
}
Please ensure the column order is correct!
Default internal values: call \code{schematize(help = TRUE)}}

\item{help}{logical.
Set to TRUE to supress execution and get help}
}
\value{
data frame with added column:
\describe{
\item{info}{character. Label of module}
}
}
\description{
\code{schematize()} assigns user-provided labels
to data, in relation on contingency.
}
\details{
Various contingencies mat belong to different categories
of experimental scheme, e.x. module of adaptation consists
of contingencies labelled as "0x0.9" or/and "0x0", whereas
the reversals stage of "0x0.3x0.9" or/and "0.3x0.9".
Defaults are set internally (call \code{schematize(help=TRUE)}),
but if the \code{info} data frame is provided the current
values get assigned to the dataset.
}
\examples{
dx <- contingencise(dx) \%>\% schematize()

dx <- contingencise(dx) \%>\% schematize(
info = data.frame(cont = 1:10,
mod = c("preadapt", rep("test",10))))

}
