% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simple}
\alias{simple}
\title{Preprocess Experiment's Files ++}
\usage{
simple(a = import(), ...)
}
\arguments{
\item{a}{dataset or call of \code{import(path-to-the-data-directory)}}

\item{...}{arguments to the \code{contingenise()}.
Each corner condition is coupled with user-specified reward access
probability, yet the the probabilities are not
directly named in data. Coding for them must be provided.
In case of problems please import the dataset with
 \code{import()} and then call \code{hint(data)} or \code{digcondition(data)}.}
}
\value{
tibble
}
\description{
\code{simple} is a wrapper simplifing
basics of the data import. Get through loading
a dataset, standardise and assign
experimental contingencies.
}
