% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking.R
\name{specify}
\alias{specify}
\title{Is That The Dataset I'm Looking For?}
\usage{
specify(A)
}
\arguments{
\item{A}{data frame}
}
\value{
list
\describe{
\item{time}{when the experiment were performed}
\item{nanimal}{number of participants}
\item{animal}{identifiers of participants}
\item{device}{list house cages used}
\item{condition}{list condition labels}
\item{temperature}{range of temperatures noted}
\item{illumination}{range of illumination noted}
\item{ifopened}{marks for door-openings (=reward access);
checks if the dataset is proper}
}
}
\description{
\code{specify()} provides basic information
about the dataset
}
\examples{
dx <- specify(dx)
}
