% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managephase.R
\name{contingencise}
\alias{contingencise}
\title{Assign Contingency of Experimental Scheme}
\usage{
contingencise(A, metadata = digform(1)$meta)
}
\arguments{
\item{A}{data frame}

\item{metadata}{named vector. Numeric vector of reward probability
assigned to corner-condition
\describe{
\item{values}{numeric values of reward probabilities}
\item{names()}{character vector of labels present in data}
}
Default: call \code{digform(version)$meta}}
}
\value{
data frame with added columns:
\describe{
\item{rp}{numeric. reward probability}
\item{contingency}{integer. number of experimental stage (=contingency)}
}
}
\description{
\code{contingencise()} assigns a contingency mark
(number of experimntal stage) to every visit in dataset
on the basis on reward probabilities.
}
\examples{

dx <- contingencise(dx)

}
