% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{import}
\alias{import}
\title{Preprocess Experiment's Files}
\usage{
import(.dir = utils::choose.dir(), .version = NULL, .list.form = NULL)
}
\arguments{
\item{.dir}{Character vector. Paths to one single directory with files}

\item{.version}{Character or integer. Version of software:
#1  {"new", alteratively: "star", "newer", 1}
#2  {"old", alteratively: "plus", "older, "proper", 2}}

\item{.list.form}{Optional list modifying defaults of \code{digform()}.
For details call \code{?digform(2)}}
}
\value{
list of data frames
}
\description{
\code{import} imports all experimental files
stored in a given directory, parses on the basis on
prespecified standards and returns a list
}
\details{
Warning: The output object (list of data frames)
can be large in size. It depends entirely on number
of files provided. Loading one experiment dataset
per time is recomended.
}
