% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simple}
\alias{simple}
\title{Preprocess Experiment's Files ++}
\usage{
simple(a = import(), ...)
}
\arguments{
\item{a}{result of \code{import()} or simply eval of it}

\item{...}{arguments to the \code{contingenise()}}
}
\value{
tibble
}
\description{
\code{simple} serves the purpose of making
welcome dummy users who do not know how
to bind a list of df into one data frame.
Presumably they also don't have a clue how to
make the furhter steps of data preprocessing
so the function is a wrapper simplifing
basics of the data import.
}
\details{
Warning: The input must be able to get
passed to dplyr::bind_rows meaning ex
names of each tibble in list must match.
}
